Improvement: refactor components, remove unwanted tests, replace dummy data
