+++ src/initialData.json
@@ -12,16 +12,7 @@
           "postedAt": "1h ago",
           "image": "https://www.telegraph.co.uk/content/dam/Travel/2019/January/greek-seaside-parga.jpg",
           "likedBy": [
-            "Diana",
-            "Sun",
-            "Livia",
-            "Marcel",
-            "Yamal",
-            "Clever",
-            "Patricia",
-            "Sua",
-            "Tao"
-            
+           
           ],
           "comments": [
             {
